EXPERIMENTAL NEW PROPOSAL FOR CONSTRAINT EDITOR:
First play around with custom proposal in ConstraintDialog, e.g. to find features in F by sub strings (F:={FeatureA, FeatureB} and input is "ureA" --> "FeatureA")) or more advanced actions like selecting a text inside the input and surround it with braces (such things). This cannot be done with standard Proposal Adapter from eclipse and a more or less working rewrite is available in this commit. However, there are some issues with line breaks inside the input.